requirements.txt
pandas>=1.1.3
ALMAFE-Lib>=0.0.12
pydantic